Moved hidden module iter out of internal
